+++ src/components/layouts/Header/index.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { AppBar, Box, Button, Stack, Typography } from '@mui/material';
+import { AppBar, Box, Stack, Typography } from '@mui/material';
 import LoginIcon from '@mui/icons-material/Login';
 import LogoutIcon from '@mui/icons-material/Logout';
 import MenuIcon from '@mui/icons-material/Menu';
 import FavoriteBorderIcon from '@mui/icons-material/FavoriteBorder';
 import SearchIcon from '@mui/icons-material/Search';
 
-import { PATH_ROOT, PATH_USER_SIGNIN } from '../../../constants/PathConstants';
+import { PATH_ROOT, PATH_USER_SIGNIN,PATH_FAVLIST } from '../../../constants/PathConstants';
 import { useDispatch, useSelector } from 'react-redux';
 import { setUser } from '../../../store/features/userSlice';
 import { Link } from 'react-router-dom';
@@ -15,6 +15,7 @@ import {
   StyledLinkButton,
   StyledSearchBox,
   StyledIconWrapperBox,
+  StyledBtn,
   StyledInputBase,
 } from './header.styled';
 import { IconButton } from '@mui/material';
@@ -34,11 +35,9 @@ export default function Header({ showSideNav }) {
         <StyledToolbar>
           {/* logo */}
           <StyledLinkButton>
-            <Typography variant="subtitle1" sx={{ flexGrow: 1 }}>
-              <Link to={PATH_ROOT} href="#">
-                Butler
-              </Link>
-            </Typography>
+            <Link to={PATH_ROOT} href="#">
+              Butler
+            </Link>
           </StyledLinkButton>
 
           {/* search */}
@@ -48,27 +47,25 @@ export default function Header({ showSideNav }) {
             </StyledIconWrapperBox>
             <StyledInputBase placeholder="Search…" inputProps={{ 'aria-label': 'search' }} />
           </StyledSearchBox>
-          <Stack direction="row" spacing={1}>
+          <Stack direction="row" justifyContent="flex-end" alignItems="center" spacing={2}>
             {/* favList */}
-            <StyledLinkButton>
-              <Typography variant="subtitle1" sx={{ flexGrow: 1 }}>
-                <Link to={PATH_ROOT} href="#">
+            <StyledBtn onClick={showSideNav} variant="contained" startIcon={<FavoriteBorderIcon color="warning" />}>
                   Favorites
-                </Link>
-              </Typography>
-            </StyledLinkButton>
+            </StyledBtn>
 
             {/* login/signup */}
             {user ? (
-              <StyledLinkButton onClick={onClickLogout}>
-                <LogoutIcon size="large" />
-              </StyledLinkButton>
+              <StyledBtn onClick={onClickLogout} variant="contained" startIcon={<LogoutIcon color="primary" />}>
+                <Typography mt={0.5} variant="subtitle2">
+                  Logout
+                </Typography>
+              </StyledBtn>
             ) : (
-              <StyledLinkButton>
+              <StyledBtn onClick={onClickLogout} variant="contained" startIcon={<LoginIcon color="primary" />}>
                 <Link to={PATH_USER_SIGNIN} href="#">
-                  <LoginIcon size="large" />
+                  Login
                 </Link>
-              </StyledLinkButton>
+              </StyledBtn>
             )}
 
             {/* sidemenu */}